requirements.txt
Flask
psycopg2-binary
Flask-SQLAlchemy
flask-cors==4.0.0